updating GitHub backend to provide header token
The current social core provides the token via a
GET parameter, and as of June or July 2020
this will be deprecated. I do not want to upgrade
social-core in the change that it breaks someones current
registry, so instead I am tweaking the backend to simply
provide the correct token via a header instead of GET param.

Signed-off-by: vsoch <vsochat@stanford.edu>